backend/requirements.txt
pydantic
pandas
pytest
dotenv
supabase
fyers-apiv3
playwright
pytz
fyers_apiv3
google-genai
fastapi